+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * a30909e - 2017-08-04: grunt tasks 
+- * 5dd67a5 - 2017-08-04: Release v9.0.60 
+- * 1735eaf - 2017-08-04: grunt tasks 
 - * d5572b7 - 2017-08-04: v9.0.45 
 - * 9ee5883 - 2017-08-04: Fixed issue #41 - downloadSubtitles() got an unexpected keyword argument &#x27;force&#x27; 
 - * 9983f42 - 2017-08-04: v9.0.44